Add files via upload
Aim of the Project:
The aim of this project is to develop and compare various deep learning models for time series forecasting using TensorFlow. The focus is on predicting future values of a dataset by training models on historical data and evaluating their performance based on Mean Absolute Error (MAE).